Some renaming and comments
